Move component tests into container directory
