+++ src/tests/mocks/MockSpotifyErrorFunctions.js
@@ -1,0 +1,7 @@
+class MockSpotifyErrorFunctions {
+  createAndSavePlaylist = async (playlist, title) => {
+    return Promise.reject({ message: 'Could not get the username.' });
+  }
+}
+
+export default MockSpotifyErrorFunctions;
