+++ src/tests/mocks/MockSpotifySuccessFunctions.js
@@ -1,0 +1,7 @@
+class MockSpotifySuccessFunctions {
+  createAndSavePlaylist = async (playlist, title) => {
+    return Promise.resolve(playlist);
+  }
+}
+
+export default MockSpotifySuccessFunctions;
